requirements.txt
django
djangorestframework
requests
Pillow